# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VPN Bot API Documentation",
        "description": "API documentation for VPN Bot Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api": {},
        "/api/v1/keys": {
            "get": {
                "tags": [
                    "Keys"
                ],
                "summary": "Список ключей пользователя",
                "operationId": "638c2f182a74fb56fbc2ecb1759ca9bc",
                "parameters": [
                    {
                        "name": "telegram_id",
                        "in": "query",
                        "description": "Telegram ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Смещение для пагинации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество записей на странице",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/KeyShort"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/keys/{keyId}": {
            "get": {
                "tags": [
                    "Keys"
                ],
                "summary": "Детали ключа",
                "operationId": "d6dce6558475bca4e975f4fd9f9f0b32",
                "parameters": [
                    {
                        "name": "telegram_id",
                        "in": "query",
                        "description": "Telegram ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Key"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/keys/{keyId}/config": {
            "get": {
                "tags": [
                    "Keys"
                ],
                "summary": "Получение конфига ключа",
                "operationId": "bb4d22071034ad39bb0b89e0150452da",
                "parameters": [
                    {
                        "name": "telegram_id",
                        "in": "query",
                        "description": "Telegram ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/keys/checkout": {
            "post": {
                "tags": [
                    "Keys"
                ],
                "summary": "Создание заказа на покупку ключа",
                "operationId": "6e26d2d1bce5fc3d40cfbef38951bef5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/KeyOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/KeyResponseDto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/keys/free-key": {
            "post": {
                "tags": [
                    "Keys"
                ],
                "summary": "Получение бесплатного ключа",
                "operationId": "b63051567c6cd7d28333af2f09de3108",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FreeKeyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "config": {
                                            "description": "Конфиг",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/keys/{keyId}/renew": {
            "post": {
                "tags": [
                    "Keys"
                ],
                "summary": "Продление ключа",
                "operationId": "ea4d9f1b06155c3258480339f895335e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RenewKeyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/KeyResponseDto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/keys/accept-payment": {
            "post": {
                "tags": [
                    "Keys"
                ],
                "summary": "(ВРЕМЕННО) Подтверждение платежа",
                "operationId": "82eda7d76e9699087fa0fbf110991f1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/KeyOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "config": {
                                            "description": "Конфиг",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/periods": {
            "get": {
                "tags": [
                    "Periods"
                ],
                "summary": "Список периодов",
                "operationId": "44d5d0bf7637b80814ccb7d738aadcf5",
                "parameters": [
                    {
                        "name": "telegram_id",
                        "in": "query",
                        "description": "Telegram ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Region"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/prices": {
            "get": {
                "tags": [
                    "Prices"
                ],
                "summary": "Список стоимости по регионам",
                "operationId": "46ce0cd9c905b7c8881d947a33f8df37",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "ID региона",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "period_id",
                        "in": "query",
                        "description": "ID периода",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Смещение для пагинации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество записей на странице",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "quantity": {
                                                        "description": "Количество ключей",
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "price": {
                                                        "type": "integer",
                                                        "example": "100"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/regions": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Список поддерживаемых регионов",
                "operationId": "473f6d445e6dba2b0178508c043c95ad",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Смещение для пагинации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество записей на странице",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Region"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Создание пользователя",
                "operationId": "ba548b97148bb5c0d8cab3bc3f128bfd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "KeyResponseDto": {
                "description": "Данные покупаемого ключа",
                "properties": {
                    "region_name": {
                        "description": "Название региона",
                        "type": "string"
                    },
                    "period_name": {
                        "description": "Название периода",
                        "type": "string"
                    },
                    "quantity": {
                        "description": "Количество ключей",
                        "type": "integer"
                    },
                    "amount": {
                        "description": "Сумма",
                        "type": "integer"
                    },
                    "payment_link": {
                        "description": "Ссылка на оплату",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "StoreUserDto": {
                "required": [
                    "name",
                    "telegram_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "telegram_id": {
                        "type": "string",
                        "example": "123456789"
                    }
                },
                "type": "object"
            },
            "FreeKeyRequest": {
                "description": "Запрос на получение бесплатного ключа",
                "properties": {
                    "telegram_id": {
                        "description": "Telegram ID пользователя",
                        "type": "string"
                    },
                    "region_id": {
                        "description": "ID региона",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RenewKeyRequest": {
                "description": "Запрос на продление ключа",
                "properties": {
                    "telegram_id": {
                        "description": "Telegram ID пользователя",
                        "type": "string"
                    },
                    "key_id": {
                        "description": "ID ключа",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "KeyOrderRequest": {
                "required": [
                    "user_id",
                    "region_id",
                    "period_id"
                ],
                "properties": {
                    "telegram_id": {
                        "description": "Telegram ID пользователя",
                        "type": "string"
                    },
                    "region_id": {
                        "description": "ID региона",
                        "type": "integer"
                    },
                    "period_id": {
                        "description": "ID периода",
                        "type": "integer"
                    },
                    "quantity": {
                        "description": "Количество ключей (от 1 до 5)",
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1,
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Key": {
                "description": "Ключ",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/KeyShort"
                    },
                    {
                        "properties": {
                            "expiration_date": {
                                "description": "Дата экспирации",
                                "type": "string",
                                "format": "date-time"
                            },
                            "amount": {
                                "description": "Стоимость продления",
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "KeyShort": {
                "description": "Ключ",
                "properties": {
                    "id": {
                        "description": "ID ключа",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название ключа",
                        "type": "string"
                    },
                    "region_name": {
                        "description": "Название региона",
                        "type": "string"
                    },
                    "region_flag": {
                        "description": "Флаг региона",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Period": {
                "description": "Период",
                "properties": {
                    "id": {
                        "description": "ID периода",
                        "type": "integer"
                    },
                    "name": {
                        "description": "ID периода",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Region": {
                "description": "Регион",
                "properties": {
                    "id": {
                        "description": "ID региона",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название региона",
                        "type": "string"
                    },
                    "code": {
                        "description": "Код региона",
                        "type": "string"
                    },
                    "flag": {
                        "description": "Флаг региона",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Keys",
            "description": "Keys"
        },
        {
            "name": "Periods",
            "description": "Periods"
        },
        {
            "name": "Prices",
            "description": "Prices"
        },
        {
            "name": "Regions",
            "description": "Regions"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}